# device_manager_service/models/machine_cycle_by_user.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from device_manager_service.models.base_model_ import Model
from device_manager_service.models.machine_cycle_by_device import MachineCycleByDevice
from device_manager_service import util

from device_manager_service.models.machine_cycle_by_device import MachineCycleByDevice  # noqa: E501

class MachineCycleByUser(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, user_id=None, cycles=None):  # noqa: E501
        """MachineCycleByUser - a model defined in OpenAPI

        :param user_id: The user_id of this MachineCycleByUser.  # noqa: E501
        :type user_id: str
        :param cycles: The cycles of this MachineCycleByUser.  # noqa: E501
        :type cycles: List[MachineCycleByDevice]
        """
        self.openapi_types = {
            'user_id': str,
            'cycles': List[MachineCycleByDevice]
        }

        self.attribute_map = {
            'user_id': 'user_id',
            'cycles': 'cycles'
        }

        self._user_id = user_id
        self._cycles = cycles

    @classmethod
    def from_dict(cls, dikt) -> 'MachineCycleByUser':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The MachineCycleByUser of this MachineCycleByUser.  # noqa: E501
        :rtype: MachineCycleByUser
        """
        return util.deserialize_model(dikt, cls)

    @property
    def user_id(self):
        """Gets the user_id of this MachineCycleByUser.


        :return: The user_id of this MachineCycleByUser.
        :rtype: str
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id):
        """Sets the user_id of this MachineCycleByUser.


        :param user_id: The user_id of this MachineCycleByUser.
        :type user_id: str
        """
        if user_id is None:
            raise ValueError("Invalid value for `user_id`, must not be `None`")  # noqa: E501
        if user_id is not None and len(user_id) > 10:
            raise ValueError("Invalid value for `user_id`, length must be less than or equal to `10`")  # noqa: E501
        if user_id is not None and len(user_id) < 10:
            raise ValueError("Invalid value for `user_id`, length must be greater than or equal to `10`")  # noqa: E501

        self._user_id = user_id

    @property
    def cycles(self):
        """Gets the cycles of this MachineCycleByUser.


        :return: The cycles of this MachineCycleByUser.
        :rtype: List[MachineCycleByDevice]
        """
        return self._cycles

    @cycles.setter
    def cycles(self, cycles):
        """Sets the cycles of this MachineCycleByUser.


        :param cycles: The cycles of this MachineCycleByUser.
        :type cycles: List[MachineCycleByDevice]
        """
        if cycles is None:
            raise ValueError("Invalid value for `cycles`, must not be `None`")  # noqa: E501

        self._cycles = cycles