# device_manager_service/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: vasco.m.campos@inesctec.pt
    name: Vasco Campos
  description: |-
    Device Manager Service OpenAPI definition.

    This service has the following functions: scan, add, delete, send commands, view measurements and state of the devices.

    Find out more: [Device Manager Service documentation](https://gitlab.inesctec.pt/cpes/european-projects/interconnect/hems/hems-documentation/-/blob/master/Microservices/Device-Manager-Service.adoc)
  title: Device Manager Service
  version: 1.0.0
servers:
- url: https://interconnect-dev.inesctec.pt/api/device
security:
- {}
- Bearer: []
...
paths:
...
  /schedule-cycle-by-device:
    description: Mock manufacturer method to schedule a cycle to a machine.
    get:
      operationId: schedule_cycle_by_device_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: serial_numbers
        required: true
        schema:
          example:
          - 1.85888214E8
          - BOSCH-WAX28EH0ES-68A40E946F87
          items:
            $ref: '#/components/schemas/SerialNumber'
          minItems: 1
          type: array
        style: form
      - explode: true
        in: query
        name: start_timestamp
        required: false
        schema:
          $ref: '#/components/schemas/Timestamp'
        style: form
      - explode: true
        in: query
        name: end_timestamp
        required: false
        schema:
          $ref: '#/components/schemas/Timestamp'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MachineCycleByDevice'
                type: array
          description: List of machine scheduled cycles per device
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Get scheduled cycles per device.
      tags:
      - Device schedules
      x-openapi-router-controller: device_manager_service.controllers.device_schedules_controller
    post:
      deprecated: true
      operationId: schedule_cycle_by_device_post
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: serial_number
        required: true
        schema:
          $ref: '#/components/schemas/SerialNumber'
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCycleRequestBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineCycle'
          description: Cycle created successfully
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Schedule new cycle for appliance (MOCK manufacturer SSA endpoint).
      tags:
      - Mock endpoints
      x-openapi-router-controller: device_manager_service.controllers.mock_endpoints_controller
  /schedule-cycle-by-user:
    description: Scheduled cycles per user.
    get:
      operationId: schedule_cycle_by_user_get
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: user_ids
        required: true
        schema:
          example:
          - userId1
          - userId2
          items:
            $ref: '#/components/schemas/UserId'
          minItems: 1
          type: array
        style: form
      - explode: true
        in: query
        name: start_timestamp
        required: false
        schema:
          $ref: '#/components/schemas/Timestamp'
        style: form
      - explode: true
        in: query
        name: end_timestamp
        required: false
        schema:
          $ref: '#/components/schemas/Timestamp'
        style: form
      - explode: true
        in: query
        name: is_cycle_optimized
        required: false
        schema:
          $ref: '#/components/schemas/Optimized'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MachineCycleByUser'
                type: array
          description: List of machine scheduled cycles by user
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Get scheduled cycles per user.
      tags:
      - Device schedules
      x-openapi-router-controller: device_manager_service.controllers.device_schedules_controller
...
components:
...
  headers:
    CorrelationId:
      explode: false
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'
      style: simple
...
  schemas:
    CorrelationId:
      format: uuid
      type: string
    UserId:
      example: 4c5e80dfjc
      maxLength: 10
      minLength: 10
      title: user_id
      type: string
    Authorization:
      format: byte
      type: string
    Timestamp:
      example: 2022-06-30T23:59:59Z
      format: date-time
      title: Timestamp
      type: string
...
    Optimized:
      enum:
      - "true"
      - "false"
      example: "false"
      title: Optimized
      type: string
...
    ScheduleCycleRequestBody:
      example:
        program: cotton
        scheduled_start_time: 2022-06-30T23:59:59Z
      properties:
        scheduled_start_time:
          example: 2022-06-30T23:59:59Z
          format: date-time
          title: earliest_start_time
          type: string
        program:
          example: cotton
          type: string
      required:
      - program
      - scheduled_start_time
      type: object
...
    SerialNumber:
      example: WPR4LLLG8NWD4
      title: serial_number
      type: string
...
    PowerUnits:
      enum:
      - W
      - kW
      title: PowerUnits
      type: string
...
    ShiftableMachineCycle:
      properties:
        sequence_id:
          example: 40A45E3C-AEDF-456A-A7F8-5A9FF3F8CD97
          title: sequence_id
          type: string
        earliest_start_time:
          example: 2022-06-30T23:59:59Z
          format: date-time
          title: Timestamp
          type: string
        latest_end_time:
          example: 2022-06-30T23:59:59Z
          format: date-time
          title: Timestamp
          type: string
        scheduled_start_time:
          example: 2022-06-30T23:59:59Z
          format: date-time
          title: Timestamp
          type: string
        expected_end_time:
          example: 2022-06-30T23:59:59Z
          format: date-time
          title: Timestamp
          type: string
        program:
          example: cotton
          type: string
        is_optimized:
          example: true
          type: boolean
        power_profile:
          items:
            $ref: '#/components/schemas/PowerProfile'
          type: array
      required:
      - expected_end_time
      - is_optimized
      - power_profile
      - program
      - scheduled_start_time
      - sequence_id
      type: object
...
    PowerProfile:
      properties:
        slot:
          example: 1
          title: slot
          type: integer
        max_power:
          example: 300
          minimum: 0
          title: max_power
          type: number
        min_power:
          example: 100
          minimum: 0
          title: min_power
          type: number
        expected_power:
          example: 200
          minimum: 0
          title: expected_power
          type: number
        power_units:
          $ref: '#/components/schemas/PowerUnits'
        duration:
          example: 4.0
          minimum: 0
          title: duration
          type: number
        duration_units:
          enum:
          - minutes
          example: minutes
          title: duration_units
          type: string
      required:
      - duration
      - duration_units
      - max_power
      - power_units
      - slot
      title: PowerProfile
      type: object
    MachineCycleByDevice:
      example:
        serial_number: WPR4LLLG8NWD4
        cycles:
        - null
        - null
      properties:
        serial_number:
          example: WPR4LLLG8NWD4
          title: serial_number
          type: string
        cycles:
          items:
            $ref: '#/components/schemas/MachineCycle'
          title: cycles
          type: array
      required:
      - cycles
      - serial_number
      title: MachineCycleByDevice
      type: object
    MachineCycleByUser:
      example:
        user_id: 4c5e80dfjc
        cycles:
        - serial_number: WPR4LLLG8NWD4
          cycles:
          - null
          - null
        - serial_number: WPR4LLLG8NWD4
          cycles:
          - null
          - null
      properties:
        user_id:
          example: 4c5e80dfjc
          maxLength: 10
          minLength: 10
          title: user_id
          type: string
        cycles:
          items:
            $ref: '#/components/schemas/MachineCycleByDevice'
          title: cycles
          type: array
      required:
      - cycles
      - user_id
      title: MachineCycleByUser
      type: object
...
    MachineCycle:
      oneOf:
      - $ref: '#/components/schemas/ShiftableMachineCycle'
      title: MachineCycle
...
    Error:
      description: Error model
      properties:
        error:
          title: error
          type: string
      required:
      - error
      title: Error
      type: object
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: device_manager_service.controllers.security_controller_.info_from_Bearer